Adds camera polish.
