+++ SkateConnect/SkateConnect/Views/Camera/CameraViewModel.swift
@@ -1,0 +1,225 @@
+//
+//  CameraViewModel.swift
+//  SkateConnect
+//
+//  Created by Konstantin Yurchenko, Jr on 9/26/24.
+//
+
+import AVFoundation
+import SwiftUI
+
+class CameraViewModel: NSObject, ObservableObject, AVCaptureFileOutputRecordingDelegate {
+    @Published var channelId: String = ""
+    @Published var isRecording = false
+    @Published var isVideoRecorded = false
+    @Published var showingPreview = false
+    @Published var showZoomHint = true
+    @Published var hasCameraAccess = true
+    @Published var isUploading = false
+    @Published var showingAlert = false
+
+    // Zoom factors for different camera modes
+    var zoomFactor: CGFloat = 1.0 // Current zoom factor
+    var wideAngleZoomFactor: CGFloat = 1.0 // Last zoom factor for wide-angle mode
+    var standardZoomFactor: CGFloat = 1.0 // Last zoom factor for standard mode
+
+    private var currentDevice: AVCaptureDevice?
+    let session = AVCaptureSession()
+    private let videoOutput = AVCaptureMovieFileOutput()
+    var videoURL: URL?
+
+    let keychainForAws = AwsKeychainStorage()
+    private let uploadManager: UploadManager
+
+    override init() {
+        uploadManager = UploadManager(keychainForAws: keychainForAws)
+        super.init()
+    }
+
+    // MARK: - Check Permissions and Setup
+    func checkPermissionsAndSetup() {
+        switch AVCaptureDevice.authorizationStatus(for: .video) {
+        case .authorized:
+            hasCameraAccess = true
+            configureSession()
+        case .notDetermined:
+            AVCaptureDevice.requestAccess(for: .video) { granted in
+                DispatchQueue.main.async {
+                    self.hasCameraAccess = granted
+                    if granted {
+                        self.configureSession()
+                    } else {
+                        print("Camera access denied.")
+                    }
+                }
+            }
+        case .denied, .restricted:
+            hasCameraAccess = false
+            print("Camera access denied or restricted.")
+        @unknown default:
+            hasCameraAccess = false
+            print("Unknown camera permission status.")
+        }
+    }
+
+    // MARK: - Configure Camera Session
+    func configureSession() {
+        session.beginConfiguration()
+
+        // Check if the camera is available
+        guard let videoDevice = AVCaptureDevice.default(for: .video) else {
+            print("Camera is not available or disabled.")
+            hasCameraAccess = false
+            return
+        }
+
+        do {
+            let videoInput = try AVCaptureDeviceInput(device: videoDevice)
+            if session.canAddInput(videoInput) {
+                session.addInput(videoInput)
+            }
+        } catch {
+            print("Error: Unable to add video input.")
+            hasCameraAccess = false
+        }
+
+        // Add audio input
+        if let audioDevice = AVCaptureDevice.default(for: .audio) {
+            do {
+                let audioInput = try AVCaptureDeviceInput(device: audioDevice)
+                if session.canAddInput(audioInput) {
+                    session.addInput(audioInput)
+                }
+            } catch {
+                print("Error: Unable to add audio input.")
+            }
+        }
+
+        // Add video output
+        if session.canAddOutput(videoOutput) {
+            session.addOutput(videoOutput)
+        }
+
+        session.commitConfiguration()
+
+        // Start the session
+        DispatchQueue.global(qos: .background).async {
+            self.session.startRunning()
+        }
+    }
+
+    // MARK: - Start Recording
+    func startRecording() {
+        let outputDirectory = FileManager.default.temporaryDirectory
+        let fileName = UUID().uuidString + ".mov"
+        videoURL = outputDirectory.appendingPathComponent(fileName)
+
+        guard let url = videoURL else { return }
+
+        videoOutput.startRecording(to: url, recordingDelegate: self)
+        isRecording = true
+    }
+
+    // MARK: - Stop Recording
+    func stopRecording() {
+        videoOutput.stopRecording()
+        isRecording = false
+    }
+
+    // MARK: - Delegate Method for Saving File
+    func fileOutput(_ output: AVCaptureFileOutput, didFinishRecordingTo outputFileURL: URL, from connections: [AVCaptureConnection], error: Error?) {
+        if let error = error {
+            print("Recording error: \(error)")
+            return
+        }
+
+        print("Video saved at: \(outputFileURL)")
+        isVideoRecorded = true
+        videoURL = outputFileURL
+    }
+
+    // MARK: - Upload Files
+    func uploadFiles(imageURL: URL) async throws {
+        isUploading = true
+
+        // Use the UploadManager to upload the image and video
+        do {
+            try await uploadManager.uploadImage(imageURL: imageURL, channelId: channelId)
+            if let videoURL = videoURL {
+                try await uploadManager.uploadVideo(videoURL: videoURL, channelId: channelId)
+            }
+        } catch {
+            print("Upload failed: \(error)")
+            throw error
+        }
+
+        isUploading = false
+        showingPreview = false
+        showingAlert = true
+    }
+
+    // MARK: - Zoom Functionality
+    func zoom(factor: CGFloat) {
+        guard let device = AVCaptureDevice.default(for: .video), device.activeFormat.videoMaxZoomFactor > factor else {
+            return
+        }
+        do {
+            try device.lockForConfiguration()
+            device.videoZoomFactor = factor
+            device.unlockForConfiguration()
+
+            // Save the zoom factor globally
+            self.zoomFactor = factor
+        } catch {
+            print("Failed to set zoom: \(error)")
+        }
+    }
+
+    // Function to switch between wide and normal lenses and restore zoom
+    func switchToWideAngle() {
+        guard let wideDevice = AVCaptureDevice.default(.builtInWideAngleCamera, for: .video, position: .back) else {
+            print("Wide angle camera not available")
+            return
+        }
+
+        // Store the current zoom factor for the standard camera
+        standardZoomFactor = zoomFactor
+
+        // Switch camera and restore last zoom level for wide-angle
+        switchCamera(to: wideDevice)
+        zoom(factor: wideAngleZoomFactor)
+    }
+
+    func switchToStandard() {
+        guard let normalDevice = AVCaptureDevice.default(.builtInTelephotoCamera, for: .video, position: .back) else {
+            print("Standard telephoto camera not available")
+            return
+        }
+
+        // Store the current zoom factor for the wide-angle camera
+        wideAngleZoomFactor = zoomFactor
+
+        // Switch camera and restore last zoom level for standard camera
+        switchCamera(to: normalDevice)
+        zoom(factor: standardZoomFactor)
+    }
+
+    private func switchCamera(to device: AVCaptureDevice) {
+        session.beginConfiguration()
+
+        // Remove all inputs
+        session.inputs.forEach { session.removeInput($0) }
+
+        do {
+            let newInput = try AVCaptureDeviceInput(device: device)
+            if session.canAddInput(newInput) {
+                session.addInput(newInput)
+                currentDevice = device
+            }
+        } catch {
+            print("Failed to switch camera: \(error)")
+        }
+
+        session.commitConfiguration()
+    }
+}
